rename versions.py to data_dict.py
